edit some styles, combine teacher and student nav into one component
